Refactor common components to designated folders and files, fix minor styling issues
