+++ frontend/src/common/components/dashboard/Dashboard.scss
@@ -1,3 +1,5 @@
+@import "../../../themes/orbcStyles.scss";
+
 .tabpanel-container {
   padding: 0 8.551vw;
   overflow: hidden;
@@ -12,6 +14,12 @@
 
 .MuiTabs-scrollButtons.Mui-disabled {
   display: none;
+}
+
+.display-tabs {
+  .MuiTab-root.Mui-selected {
+    color: $bc-text-links-blue;
+  }
 }
 
 @media screen and (max-width: 768px) {
