+++ frontend/src/common/components/table/options/CSVOptions.scss
@@ -1,0 +1,12 @@
+.options-container {
+  display: flex;
+  padding: 7px 0 7px 0;
+}
+
+.csv-container {
+  margin-left: auto;
+}
+
+.trash {
+  padding: 0px 30px;
+}
