+++ frontend/src/common/components/table/options/Filter.scss
@@ -1,0 +1,4 @@
+.filter-btn {
+  padding: 0 1rem;
+  margin: 0 1.25rem;
+}
