+++ frontend/src/common/components/table/options/Filter.tsx
@@ -2,17 +2,19 @@ import { Button } from "@mui/material";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFilter } from "@fortawesome/free-solid-svg-icons";
 
+import "./Filter.scss";
+
 export const Filter = () => {
   return (
     <>
       <Button
+        className="filter-btn"
         aria-label="filter"
         id="filter-button"
         variant="contained"
         color="tertiary"
         aria-haspopup="true"
         startIcon={<FontAwesomeIcon icon={faFilter} />}
-        sx={{ padding: "0px 16px", margin: "0px 20px" }}
         disabled={true}
       >
         Filter
