+++ frontend/src/common/components/table/options/Trash.scss
@@ -1,0 +1,3 @@
+.trash-btn {
+  padding: 0.625rem 0.625rem;
+}
