+++ frontend/src/features/manageVehicles/components/list/List.tsx
@@ -1,3 +1,8 @@
+import { RowSelectionState } from "@tanstack/table-core";
+import { Box, IconButton, Tooltip } from "@mui/material";
+import { Delete, Edit, ContentCopy } from "@mui/icons-material";
+import { UseQueryResult } from "@tanstack/react-query";
+import { useNavigate } from "react-router-dom";
 import {
   memo,
   useCallback,
@@ -6,35 +11,32 @@ import {
   useMemo,
   useState,
 } from "react";
+
 import MaterialReactTable, {
   MRT_ColumnDef,
   MRT_GlobalFilterTextField,
   MRT_Row,
   MRT_TableInstance,
 } from "material-react-table";
-import { RowSelectionState } from "@tanstack/table-core";
+
 import "./List.scss";
-import { Box, IconButton, Tooltip } from "@mui/material";
+import { Filter } from "../../../../common/components/table/options/Filter";
+import { Trash } from "../../../../common/components/table/options/Trash";
+import { CSVOptions } from "../../../../common/components/table/options/CSVOptions";
+import { DeleteConfirmationDialog } from "../../../../common/components/dialog/DeleteConfirmationDialog";
+import { BC_COLOURS } from "../../../../themes/bcGovStyles";
+import { PowerUnitColumnDefinition, TrailerColumnDefinition } from "./Columns";
+import { deleteVehicles } from "../../apiManager/vehiclesAPI";
+import { SnackBarContext } from "../../../../App";
+import { MANAGE_VEHICLES } from "../../../../routes/constants";
+import { DoesUserHaveRoleWithContext } from "../../../../common/authentication/util";
+import { ROLES } from "../../../../common/authentication/types";
 import {
   VehicleTypes,
   VehicleTypesAsString,
   PowerUnit,
   Trailer,
 } from "../../types/managevehicles";
-import { Filter } from "../options/Filter";
-import { Trash } from "../options/Trash";
-import { CSVOptions } from "../options/CSVOptions";
-import { Delete, Edit, ContentCopy } from "@mui/icons-material";
-import { BC_COLOURS } from "../../../../themes/bcGovStyles";
-import { UseQueryResult } from "@tanstack/react-query";
-import { PowerUnitColumnDefinition, TrailerColumnDefinition } from "./Columns";
-import { deleteVehicles } from "../../apiManager/vehiclesAPI";
-import DeleteConfirmationDialog from "./ConfirmationDialog";
-import { useNavigate } from "react-router-dom";
-import { SnackBarContext } from "../../../../App";
-import { MANAGE_VEHICLES } from "../../../../routes/constants";
-import { DoesUserHaveRoleWithContext } from "../../../../common/authentication/util";
-import { ROLES } from "../../../../common/authentication/types";
 
 /**
  * Dynamically set the column based on vehicle type
