--- frontend/src/features/manageVehicles/components/options/Options.scss
@@ -1,12 +1,0 @@
-.options-container {
-    display: flex;
-    padding: 7px 0 7px 0;
-}
-
-.csv-container {
-    margin-left: auto;
-}
-
-.trash {
-    padding: 0px 30px;
-}
