+++ frontend/src/features/permits/components/list/Columns.tsx
@@ -1,6 +1,7 @@
 import { MRT_ColumnDef } from "material-react-table";
-import { ApplicationInProgress } from "../../types/application";
 import { Link } from "react-router-dom";
+
+import { ApplicationInProgress } from "../../types/application";
 
 export const ApplicationInProgressColumnDefinition: MRT_ColumnDef<ApplicationInProgress>[] = [
   {
@@ -8,7 +9,14 @@ export const ApplicationInProgressColumnDefinition: MRT_ColumnDef<ApplicationInP
     header: "Application #",
     accessorFn: (row) => row.applicationNumber,
     Cell: (props: {cell: any, row: any}) => {
-      return <Link to={`/applications/${props.row.original.permitId}`}>{props.cell.getValue()}</Link>
+      return (
+        <Link 
+          to={`/applications/${props.row.original.permitId}`}
+          className="column-link column-link--application-details"
+        >
+          {props.cell.getValue()}
+        </Link>
+      );
     },
 
   },
