+++ frontend/src/features/permits/components/list/List.scss
