+++ frontend/src/features/permits/components/permit-list/BasePermitList.tsx
@@ -1,16 +1,16 @@
 import { Box } from "@mui/material";
+import { useCallback, useContext, useEffect } from "react";
 import { UseQueryResult } from "@tanstack/react-query";
 import MaterialReactTable, {
   MRT_GlobalFilterTextField,
   MRT_Row,
   MRT_TableInstance
 } from "material-react-table";
-import { useCallback, useContext, useEffect } from "react";
 
+import "../list/List.scss";
 import { SnackBarContext } from "../../../../App";
 import { NoRecordsFound } from "../../../../common/components/table/NoRecordsFound";
 import { BC_COLOURS } from "../../../../themes/bcGovStyles";
-import "../../../manageVehicles/components/list/List.scss";
 import { Permit } from "../../types/permit";
 import { PermitsColumnDefinition } from "./Columns";
 import { PermitRowOptions } from "./PermitRowOptions";
